pyproject.toml
[project]
name = "ad-synth-ai"
version = "0.1.0"
description = "Multi-agent AI platform for advertising synthesis"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "google-adk>=1.0",
  "google-genai>=1.0",
  "google-cloud-aiplatform>=1.60",
  "sqlalchemy[asyncio]>=2.0",
  "aiosqlite>=0.20",
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "bcrypt>=4.0",
  "python-jose[cryptography]>=3.3",
  "python-multipart>=0.0.9",
  "httpx>=0.27",
  "praw>=7.7",
  "google-api-python-client>=2.100",
  "tweepy>=4.14",
  "sqlite-vec>=0.1",
  "tenacity>=8.3",
  "python-dotenv>=1.0",
]

[dependency-groups]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--log-file=logs/pytest.log --log-file-level=DEBUG"
markers = [
  "unit: fast, fully mocked — no external services",
  "module: mocked external services, real in-memory DB",
  "system: end-to-end with real Gemini LLM (requires GOOGLE_API_KEY)",
  "connection: live API connectivity checks (auto-skipped if key is absent)",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend", "tools"]
